Move shared terriajs-server gulp task to separate file
